extract a base test to share with the ScalaDocOptions
